Automatic CV and Cover Letter with API/requirements.txt
openai>=1.0.0
python-docx>=0.8.11
beautifulsoup4>=4.9.3
requests>=2.25.1
python-dotenv>=0.19.0 